♻️ Move mocks into tests folder
