+++ packages/usehooks-ts/src/useDarkMode/useDarkMode.test.ts
@@ -1,6 +1,6 @@
 import { act, renderHook } from '@testing-library/react'
 
-import { mockMatchMedia, mockStorage } from '../../mocks'
+import { mockMatchMedia, mockStorage } from '../../tests/mocks'
 import { useDarkMode } from './useDarkMode'
 
 mockStorage('localStorage')
