+++ packages/usehooks-ts/src/useLocalStorage/useLocalStorage.test.ts
@@ -1,6 +1,6 @@
 import { act, renderHook } from '@testing-library/react'
 
-import { mockStorage } from '../../mocks'
+import { mockStorage } from '../../tests/mocks'
 import { useLocalStorage } from './useLocalStorage'
 
 mockStorage('localStorage')
