+++ packages/usehooks-ts/tests/mocks.ts
